requirements.txt
Package                   Version
------------------------- -----------
accelerate                1.12.0
aiohappyeyeballs          2.6.1
aiohttp                   3.13.2
aiosignal                 1.4.0
altair                    6.0.0
annotated-types           0.7.0
anyio                     4.12.0
attrs                     25.4.0
bitsandbytes              0.48.2
blinker                   1.9.0
cachetools                6.2.2
certifi                   2025.11.12
cffi                      2.0.0
charset-normalizer        3.4.4
click                     8.3.1
colorama                  0.4.6
cryptography              46.0.3
dataclasses-json          0.6.7
distro                    1.9.0
elastic-transport         8.17.1
elasticsearch             8.19.2
faiss-cpu                 1.13.1
filelock                  3.20.0
frozenlist                1.8.0
fsspec                    2025.12.0
gitdb                     4.0.12
GitPython                 3.1.45
greenlet                  3.3.0
h11                       0.16.0
httpcore                  1.0.9
httpx                     0.28.1
httpx-sse                 0.4.3
huggingface-hub           0.36.0
idna                      3.11
Jinja2                    3.1.6
jiter                     0.12.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
langchain                 0.3.12
langchain-community       0.3.12
langchain-core            0.3.63
langchain-elasticsearch   0.3.1
langchain-openai          0.2.6
langchain-text-splitters  0.3.3
langgraph-checkpoint      3.0.1
langgraph-sdk             0.2.12
langsmith                 0.2.11
MarkupSafe                3.0.3
marshmallow               3.26.1
mpmath                    1.3.0
multidict                 6.7.0
mypy_extensions           1.1.0
narwhals                  2.13.0
networkx                  3.6
numpy                     1.26.4
openai                    1.109.1
orjson                    3.11.4
ormsgpack                 1.12.0
packaging                 24.2
pandas                    2.3.3
pdfminer.six              20251107
pdfplumber                0.11.8
pillow                    12.0.0
pip                       25.3
propcache                 0.4.1
protobuf                  6.33.2
psutil                    7.1.3
pyarrow                   22.0.0
pycparser                 2.23
pydantic                  2.12.5
pydantic_core             2.41.5
pydantic-settings         2.12.0
pydeck                    0.9.1
pypdf                     6.4.0
pypdfium2                 5.1.0
python-dateutil           2.9.0.post0
python-dotenv             1.2.1
pytz                      2025.2
PyYAML                    6.0.3
referencing               0.37.0
regex                     2025.11.3
requests                  2.32.5
requests-toolbelt         1.0.0
rpds-py                   0.30.0
safetensors               0.7.0
setuptools                80.9.0
simsimd                   6.5.3
six                       1.17.0
smmap                     5.0.2
sniffio                   1.3.1
SQLAlchemy                2.0.44
streamlit                 1.52.1
sympy                     1.14.0
tenacity                  9.1.2
tiktoken                  0.12.0
tokenizers                0.22.1
toml                      0.10.2
torch                     2.9.1
tornado                   6.5.2
tqdm                      4.67.1
typing_extensions         4.15.0
typing-inspect            0.9.0
typing-inspection         0.4.2
tzdata                    2025.2
urllib3                   2.5.0
uuid_utils                0.12.0
watchdog                  6.0.0
wheel                     0.45.1
xxhash                    3.6.0
yarl                      1.22.0
zstandard                 0.25.0
